+++ apps/api/src/app/app-mongoose.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { mongoConnectionString } from './core/';
+import { mongoConnectionString } from './common';
 
 @Module({
   imports: [MongooseModule.forRoot(mongoConnectionString())],
