+++ apps/api/src/app/common/options/mongo-client.options.ts
