+++ apps/api/src/app/common/services/index.ts
@@ -1,3 +1,5 @@
+export * from './settings.service';
+
 export * from './pagination.service';
 
 export * from './grpc-error.service';
