+++ apps/api/src/app/common/services/settings.service.ts
