--- apps/api/src/app/core/core.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ConfigModule } from '@nestjs/config';
-import { PaginationService, GrpcErrorService } from './services';
-
-@Module({
-  imports: [ConfigModule.forRoot()],
-  providers: [PaginationService, GrpcErrorService],
-})
-export class CoreModule {}
