--- apps/api/src/app/core/index.ts
@@ -1,7 +1,0 @@
-export * from './core.module';
-
-export * from './options';
-
-export * from './proto';
-
-export * from './services';
