+++ apps/api/src/app/modules/applications/applications.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
-import { PaginationService } from '../../core';
-import { SharedModule } from '../../shared';
+import { CommonModule } from '../../common/common.module';
 import { ApplicationsMongooseModule } from './applications-mongoose.module';
 import { ApplicationsController } from './applications.controller';
 import { ApplicationsService } from './applications.service';
 
 @Module({
-  imports: [SharedModule, ApplicationsMongooseModule],
+  imports: [CommonModule, ApplicationsMongooseModule],
   controllers: [ApplicationsController],
-  providers: [ApplicationsService, PaginationService],
+  providers: [ApplicationsService],
 })
 export class ApplicationsModule {}
