+++ apps/api/src/app/modules/sessions/sessions.module.ts
@@ -1,11 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ClientsModule } from '@nestjs/microservices';
-import {
-  grpcClientOptions,
-  GrpcErrorService,
-  PaginationService,
-} from '../../core';
-import { SharedModule } from '../../shared/';
+import { grpcClientOptions } from '../../common';
+import { CommonModule } from '../../common/common.module';
 import { SessionsMongooseModule } from './sessions-mongoose.module';
 import { SessionsController } from './sessions.controller';
 import { SessionsService } from './sessions.service';
@@ -15,7 +11,7 @@ import { SessionsService } from './sessions.service';
  */
 @Module({
   imports: [
-    SharedModule,
+    CommonModule,
     SessionsMongooseModule,
     ClientsModule.register([
       {
@@ -25,6 +21,6 @@ import { SessionsService } from './sessions.service';
     ]),
   ],
   controllers: [SessionsController],
-  providers: [SessionsService, PaginationService, GrpcErrorService],
+  providers: [SessionsService],
 })
 export class SessionsModule {}
