+++ apps/api/src/app/modules/sessions/sessions.service.ts
@@ -14,8 +14,7 @@ import {
 import { ClientGrpc } from '@nestjs/microservices';
 import { InjectConnection, InjectModel } from '@nestjs/mongoose';
 import { Connection, Model } from 'mongoose';
-import { PaginationService, Submitter } from '../../core';
-import { SettingsService } from '../../shared';
+import { PaginationService, SettingsService, Submitter } from '../../common';
 import { Task, TaskDocument } from '../tasks/schemas';
 import { Session, SessionDocument } from './schemas';
