--- apps/api/src/app/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './shared.module';
-
-export * from './services';
