--- apps/api/src/app/shared/services/index.ts
@@ -1,1 +1,0 @@
-export * from './settings.service';
