--- apps/api/src/app/shared/shared.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SettingsService } from './services';
-
-@Module({
-  providers: [SettingsService],
-  exports: [SettingsService],
-})
-export class SharedModule {}
